gril: Use unsigned int instead of gsize
Sizeof gsize changes depending on 32/64 bit system.  However, in reality
it is always 4 bytes.

+++ gril/gril.h
@@ -44,7 +44,7 @@ typedef struct _GRil GRil;
  */
 struct ril_msg {
 	gchar *buf;
-	gsize buf_len;
+	unsigned int buf_len;
 	gboolean unsolicited;
 	int req;
 	int serial_no;
